migrate patch from ucs 1.3-2 management scope
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/trunk/ucs@63 1b283449-9f2f-0410-b571-8eb228e1a901

+++ base/univention-installer/scripts/50_packages.sh
@@ -137,8 +137,13 @@ fi
 
 links=`echo $packages | grep links-ssl`
 if [ -z "\$links" ]; then
+	# apache needs a www-browser package
+	res=\`apt-get remove -s links-ssl elinks lynx | grep apache\`
+	if [ -z "\$res" ]; then
 		$PIPE apt-get remove links-ssl elinks lynx
 		$PIPE dpkg --configure -a
+	fi
+
 fi
 
 if [ -n "$extra_packages" ]; then
